refactor: UserRepository 위치 변경
